requirements.txt
streamlit==1.29.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
openai==1.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
reportlab==4.0.7